requirements.txt
Flask
Flask-SQLAlchemy
pytesseract
Pillow
pdf2image
pandas
openpyxl